refactor(app): split events and apply clusters out of app.rs
Pure relocations, no behavior or API change — `app.rs` shrinks 478 → 288
lines by extracting two thematic clusters into sibling `impl App` files:

- events.rs: winit input handlers (on_keyboard_input,
  execute_app_command, on_mouse_wheel, on_cursor_moved, on_mouse_input)
  plus the free-fn handle_mouse_press and the font-size clamp constants.

- apply.rs: the apply_font_size / apply_font_metrics / apply_scale_factor
  / apply_window_padding cluster that propagates a settings change into
  the renderer and reflows the PTY.

app.rs keeps the App struct, lifecycle (new, ws_mut, mark_dirty), the
frame-loop pair (draw, tick_blink), the ApplicationHandler impl, and the
small helpers (initial_window_size_from_env, vt_shape_from_config,
physical_window_padding). `keybinds` field is now pub(crate) so events.rs
can read it; apply_font_size is pub(crate) so execute_app_command in
events.rs can call it.
